pyproject.toml
[tool.poetry]
name = "paper-translation-in-notion"
version = "0.1.0"
description = ""
authors = ["vodoomi <abe.mikito@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
numpy = "1.26.4"
google-generativeai = "^0.8.4"
notion-client = "^2.3.0"
mistralai = "^1.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
